refactor: move types into their proper module files
