+++ packages/account-tree-controller/src/types.ts
@@ -1,7 +1,5 @@
-import type { AccountWalletCategory } from '@metamask/account-api';
 import type { AccountGroupId, AccountWalletId } from '@metamask/account-api';
 import type {
-  AccountId,
   AccountsControllerAccountAddedEvent,
   AccountsControllerAccountRemovedEvent,
   AccountsControllerGetAccountAction,
@@ -15,81 +13,14 @@ import {
   type ControllerStateChangeEvent,
   type RestrictedMessenger,
 } from '@metamask/base-controller';
-import type { EntropySourceId } from '@metamask/keyring-api';
-import type {
-  KeyringControllerGetStateAction,
-  KeyringTypes,
-} from '@metamask/keyring-controller';
+import type { KeyringControllerGetStateAction } from '@metamask/keyring-controller';
 import type { GetSnap as SnapControllerGetSnap } from '@metamask/snaps-controllers';
-import type { SnapId } from '@metamask/snaps-sdk';
 
 import type {
   AccountTreeController,
   controllerName,
 } from './AccountTreeController';
-
-/**
- * Account wallet metadata for the "entropy" wallet category.
- */
-export type AccountWalletEntropyMetadata = {
-  type: AccountWalletCategory.Entropy;
-  entropy: {
-    id: EntropySourceId;
-    index: number;
-  };
-};
-
-/**
- * Account wallet metadata for the "snap" wallet category.
- */
-export type AccountWalletSnapMetadata = {
-  type: AccountWalletCategory.Snap;
-  snap: {
-    id: SnapId;
-  };
-};
-
-/**
- * Account wallet metadata for the "keyring" wallet category.
- */
-export type AccountWalletKeyringMetadata = {
-  type: AccountWalletCategory.Keyring;
-  keyring: {
-    type: KeyringTypes;
-  };
-};
-
-/**
- * Account wallet metadata for the "keyring" wallet category.
- */
-export type AccountWalletCategoryMetadata =
-  | AccountWalletEntropyMetadata
-  | AccountWalletSnapMetadata
-  | AccountWalletKeyringMetadata;
-
-export type AccountWalletMetadata = {
-  name: string;
-} & AccountWalletCategoryMetadata;
-
-export type AccountGroupMetadata = {
-  name: string;
-};
-
-export type AccountGroupObject = {
-  id: AccountGroupId;
-  // Blockchain Accounts:
-  accounts: AccountId[];
-  metadata: AccountGroupMetadata;
-};
-
-export type AccountWalletObject = {
-  id: AccountWalletId;
-  // Account groups OR Multichain accounts (once available).
-  groups: {
-    [groupId: AccountGroupId]: AccountGroupObject;
-  };
-  metadata: AccountWalletMetadata;
-};
+import type { AccountWalletObject } from './wallet';
 
 export type AccountTreeControllerState = {
   accountTree: {
